+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes will be documented in this file. The format is based on
 ## [Unreleased]
 ### Added
 - Support for recent Steam backups compressed with zstd.
+- `--manifest-dir` argument to `tev backup verify`, to optionally verify
+  matching cached manifest files.
+- `--fast` flag to `tev backup verify`, to skip verification of individual chunk
+  SHA-1 digests (but check everything else).
+
+### Changed
+- `tev backup verify` can now take multiple path arguments to verify multiple
+  backups in sequence.
 
 ## [0.2.0] - 2025-01-05
 ### Added